pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simlearn"
version = "0.1.0"
description = "Student Response Simulation Framework for educational data"
readme = "README.md"
requires-python = ">=3.8,<4.0"
license = { file = "LICENSE" }
authors = [
  { name="Author Name", email="author@example.com" },
]
dependencies = [
  "numpy>=1.21",
  "pandas>=1.3",
  "pydantic>=1.10",
  "typer[all]>=0.6",
  "PyYAML>=5.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=6.0",
  "pytest-cov",
  "pre-commit",
  "black",
  "isort",
  "ruff",
  "mypy",
  "types-PyYAML",
]
docs = [
  "sphinx>=4.0",
  "myst-parser",
]
numba = ["numba>=0.55"]

[project.scripts]
simulate = "simlearn.cli:main"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120

[tool.mypy]
files = ["src"]
ignore_missing_imports = true
strict = false
check_untyped_defs = false
